requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI Models - Updated for better compatibility
openai-whisper>=20231117  # Audio transcription (Whisper)
deepface>=0.0.79
tf-keras>=2.15.0  # Required for DeepFace/TensorFlow compatibility
opencv-python>=4.8.0
spacy>=3.7.0
transformers>=4.35.0
huggingface-hub>=0.23.0
sentencepiece>=0.1.99
nltk>=3.8.0
mtcnn>=0.1.1  # Modern face detection (Phase 1 accuracy improvement)
webrtcvad-wheels>=2.0.10.post2  # Prebuilt VAD wheels for Windows/Python 3.13 compatibility
scikit-learn>=1.3.0  # ML pattern recognition (Phase 3 accuracy improvement)
scipy>=1.11.0  # Scientific computing for acoustic analysis (Phase 3)

# Audio Processing
librosa>=0.10.0,<0.11.0  # Pin to avoid numba compatibility issues
soundfile>=0.12.0
numba>=0.56.0,<0.57.0   # Compatible with librosa 0.10.x

# Deep Learning - Compatible versions (updated for current PyTorch versions)
torch>=2.0.0
torchaudio>=2.0.0

# Data Processing - Compatible versions
numpy>=1.21.0,<1.25.0   # Compatible with librosa and numba
pandas>=1.5.0
pillow>=9.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.2

# Video Processing (Python-based, no FFmpeg required)
moviepy>=2.2.0
imageio-ffmpeg>=0.6.0


# Authentication & Security - Compatible versions
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
email-validator>=2.0.0  # Required for Pydantic EmailStr validation

# Database - Compatible versions
sqlalchemy>=1.4.0
alembic>=1.8.0

# API Integrations
openai>=1.3.0

# Advanced Features (optional - uncomment if needed)
# websockets>=11.0.0
# celery>=5.3.0
# redis>=5.0.0

# Development & Deployment
pytest>=7.4.0
pytest-asyncio>=0.21.0
docker>=6.1.0
gunicorn>=21.2.0

# PDF/Word Report Generation
reportlab>=4.0.0
python-docx>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Optional OpenAI API (uncomment if using OpenAI features)
# openai==1.3.0
